moved ITrackable and ISoftDelete interfaces to EntityModels namespace
