+++ Fabric.Identity.API/Persistence/SqlServer/EntityModels/ApiResource.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Fabric.Identity.API.Persistence.SqlServer.Entities;
 
 namespace Fabric.Identity.API.Persistence.SqlServer.EntityModels
 {
