+++ Fabric.Identity.API/Persistence/SqlServer/EntityModels/ISoftDelete.cs
@@ -1,4 +1,4 @@
-﻿namespace Fabric.Identity.API.Persistence.SqlServer.Entities
+﻿namespace Fabric.Identity.API.Persistence.SqlServer.EntityModels
 {
     public interface ISoftDelete
     {
