+++ Fabric.Identity.API/Persistence/SqlServer/EntityModels/ITrackable.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Fabric.Identity.API.Persistence.SqlServer.Entities
+namespace Fabric.Identity.API.Persistence.SqlServer.EntityModels
 {
     public interface ITrackable
     {
